Fixup First version of crypto context
The commit moves the `short_circuit_sig_size` and `short_circuit_sign`
functions to t_cose_util, so that is accessible to the test crypto
adapter.
