+++ src/t_cose_util.h
@@ -139,6 +139,67 @@ enum t_cose_err_t create_tbs_hash(int32_t                     cose_algorithm_id,
  *
  */
 struct q_useful_buf_c get_short_circuit_kid(void);
+
+/**
+ * \brief Calculate the size of a short-circuit signature
+ *
+ * \param[in]  cose_algorithm_id Algorithm ID.
+ * \param[out] sig_size          The size of the signature
+ *
+ * \return This returns one of the error codes defined by \ref t_cose_err_t.
+ *
+ * This returns the size of the short-circuit signature for this
+ * cose_algorithm_id
+ */
+enum t_cose_err_t
+short_circuit_sig_size(int32_t            cose_algorithm_id,
+                       size_t            *sig_size);
+
+
+/**
+ * \brief Create a short-circuit signature
+ *
+ * \param[in] cose_algorithm_id Algorithm ID. This is used only to make
+ *                              the short-circuit signature the same size
+ *                              as the real signature would be for the
+ *                              particular algorithm.
+ * \param[in] hash_to_sign      The bytes to sign. Typically, a hash of
+ *                              a payload.
+ * \param[in] signature_buffer  Pointer and length of buffer into which
+ *                              the resulting signature is put.
+ * \param[in] signature         Pointer and length of the signature
+ *                              returned.
+ *
+ * \return This returns one of the error codes defined by \ref t_cose_err_t.
+ *
+ * This creates the short-circuit signature that is a concatenation of
+ * hashes up to the expected size of the signature. This is a test
+ * mode only has it has no security value. This is retained in
+ * commercial production code as a useful test or demo that can run
+ * even if key material is not set up or accessible.
+ */
+enum t_cose_err_t
+short_circuit_sign(int32_t               cose_algorithm_id,
+                   struct q_useful_buf_c hash_to_sign,
+                   struct q_useful_buf   signature_buffer,
+                   struct q_useful_buf_c *signature);
+
+
+/**
+ * \brief Verify a short-circuit signature
+ *
+ * \param[in] hash_to_verify  Pointer and length of hash to verify.
+ * \param[in] signature       Pointer and length of signature.
+ *
+ * \return This returns one of the error codes defined by \ref
+ *         t_cose_err_t.
+ *
+ * See t_cose_sign1_sign_init() for description of the short-circuit
+ * signature.
+ */
+enum t_cose_err_t
+t_cose_crypto_short_circuit_verify(struct q_useful_buf_c hash_to_verify,
+                                   struct q_useful_buf_c signature);
 #endif
 
 #ifdef __cplusplus
